test downloader w/ + w/o directory

+++ wqio/tests/test_datasets.py
@@ -10,11 +10,18 @@ from wqio import datasets
 
 @network(url='https://github.com/Geosyntec/water-quality-dataset')
 @pytest.mark.parametrize('fname', ['bmpdata', 'nsqd'])
-def test_download(fname):
-    tmpdir = tempfile.mkdtemp()
+@pytest.mark.parametrize('use_tmpdir', [True, False])
+def test_download(fname, use_tmpdir):
+    if use_tmpdir:
+        tmpdir = tempfile.mkdtemp()
+    else:
+        tmpdir = None
+
     fname = datasets.download(fname, data_dir=tmpdir)
     assert os.path.exists(fname)
     yield
+
     path, _file = os.path.split(fname)
     os.remove(fname)
-    shutil.rmtree(tmpdir)
+    if tmpdir is not None:
+        shutil.rmtree(tmpdir)
